Addition of libraries and Test codes
Added the MCP_CAN libraries
Added the Roboclaw.cpp and Roboclaw.h libraries
Created a Test_Bed.cpp code to test out functionality of board
Added SPI.h and SPI.cpp for SPI communication libraries
Sketch.cpp was the original file used to test out codes. At this point
it is more of a note file for saved functionality and should be added
to a git ignore file.